LayoutTests/fast/repaint/drop-shadow-box-change-expected.txt

Test repaint rects when a box with drop-shadow changes size

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS layerTreeAsText.indexOf('(rect 192 192 206 156)') > -1 is true
PASS successfullyParsed is true

TEST COMPLETE
